fix(BFormInput,BFormTextArea): Deprecate trim, lazy, number props in favor of modifiers (#2066)
