+++ packages/bootstrap-vue-next/src/components/BFormInput/BFormInput.vue
@@ -27,6 +27,7 @@
 <script setup lang="ts">
 import {computed, ref} from 'vue'
 import {useDefaults, useFormInput, useStateClass} from '../../composables'
+import {normalizeInput} from '../../utils'
 import type {BFormInputProps, Numberish} from '../../types'
 
 const _props = withDefaults(defineProps<BFormInputProps>(), {
@@ -44,39 +45,23 @@ const _props = withDefaults(defineProps<BFormInputProps>(), {
   form: undefined,
   formatter: undefined,
   id: undefined,
-  lazy: false,
   lazyFormatter: false,
   list: undefined,
   modelValue: '',
   name: undefined,
-  number: false,
   placeholder: undefined,
   plaintext: false,
   readonly: false,
   required: false,
   size: undefined,
   state: null,
-  trim: false,
   // End CommonInputProps
 })
 const props = useDefaults(_props, 'BFormInput')
 
 const [modelValue, modelModifiers] = defineModel<Numberish | null, 'trim' | 'lazy' | 'number'>({
   default: '',
-  set: (v) => {
-    if (v === null) return
-    let update = v
-    if (modelModifiers.trim) update = update.toString().trim()
-    if (
-      (modelModifiers.number || props.type === 'number') &&
-      typeof update === 'string' &&
-      update !== ''
-    ) {
-      const parsed = Number.parseFloat(update)
-      update = Number.isNaN(parsed) ? update : parsed
-    }
-    return update
-  },
+  set: (v) => normalizeInput(v, modelModifiers),
 })
 
 const {input, computedId, computedAriaInvalid, onInput, onChange, onBlur, focus, blur} =
