+++ packages/bootstrap-vue-next/src/utils/index.ts
@@ -11,6 +11,7 @@ export {default as getSlotElements} from './getSlotElements'
 export {default as getTableFieldHeadLabel} from './getTableFieldHeadLabel'
 export {default as isLink} from './isLink'
 export * from './keys'
+export {default as normalizeInput} from './normalizeInput'
 export {default as normalizeSlot} from './normalizeSlot'
 export * from './object'
 export {default as parseActiveImports} from './parseActiveImports'
